+++ release-testing/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog 刀 DAO
 # Day Ahead Optimizer
+# 2026.03.0.rc1
+- Made optional battery settings "bat_to_dc max power" and "dc_to_bat max power" flex-setting (feature request by @DaBit)
+- Reduce power during charging at high soc and during discharging at low soc (feature requests form @bartzzz and @arjenhiemstra)
+- Made check 'optimal lower level" lower as "lower limit" (feature request of @mistral2)
+
 # 2026.02.2.rc4
 - Fix error in calculating heating window boiler
 - Update python module pandas 3.0.0 -> 3.0.1